requirements.txt
pyperclip==1.8.2
PyQt5==5.15.4
Pillow==8.2.0
pix2tex==0.5.2